requirements.txt
streamlit==1.32.2
tensorflow==2.15.0
pandas==2.2.1
numpy==1.26.4
matplotlib==3.8.3
scikit-learn==1.4.1.post1
Pillow==10.2.0
